# packages/shared/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NextQuest API",
    "description": "API de gestion de ludotheque videoludique. Authentification via JWT Bearer ou cookie HttpOnly.",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token JWT renvoye par /auth/login, /auth/register ou /auth/refresh."
      }
    },
...
  },
  "paths": {
...
    "/api/games/upcoming": {
      "get": {
        "summary": "Lister les jeux a venir (date de sortie future)",
        "tags": [
          "Games"
        ],
        "parameters": [
          {
            "schema": {
              "default": 20,
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "default": 0,
              "type": "integer",
              "minimum": 0,
              "maximum": 9007199254740991
            },
            "in": "query",
            "name": "offset",
            "required": false
          },
          {
            "schema": {
              "default": "hype",
              "type": "string",
              "enum": [
                "hype",
                "date"
              ]
            },
            "in": "query",
            "name": "sort",
            "required": false
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/games/igdb/{igdbId}": {
      "get": {
        "summary": "Detail d'un jeu par son IGDB ID",
        "tags": [
          "Games"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 0,
              "exclusiveMinimum": true,
              "maximum": 9007199254740991
            },
            "in": "path",
            "name": "igdbId",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/users/me/library/enrich": {
      "post": {
        "summary": "Enrichir (IGDB) la bibliotheque de l'utilisateur",
        "tags": [
          "Games"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/admin/games/enrich": {
      "post": {
        "summary": "Enrichir (IGDB) tout le catalogue",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
...
}